train_model/requirements.txt
# Ollama RAG 系統所需依賴套件

# 核心 AI/ML 套件
ollama
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0

# 向量資料庫
chromadb>=0.4.0
faiss-cpu>=1.7.4

# 資料處理
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# 文本處理
nltk>=3.8
spacy>=3.6.0
jieba>=0.42.1  # 中文分詞

# 網頁和API
requests>=2.31.0
httpx>=0.25.0

# 設定管理
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 日誌和監控
loguru>=0.7.0
tqdm>=4.65.0

# 資料視覺化
matplotlib>=3.7.0
seaborn>=0.12.0

# 其他工具
pyyaml>=6.0
colorama>=0.4.6
rich>=13.0.0